requirements.txt
annotated-doc         0.0.4
annotated-types       0.7.0
anyio                 4.12.1
blis                  1.3.3
catalogue             2.0.10
certifi               2026.1.4
charset-normalizer    3.4.4
click                 8.3.1
cloudpathlib          0.23.0
colorama              0.4.6
confection            0.1.5
cymem                 2.0.13
distro                1.9.0
en_core_web_sm        3.8.0
fastapi               0.128.8
filelock              3.20.3
fsspec                2026.2.0
h11                   0.16.0
hf-xet                1.2.0
httpcore              1.0.9
httpx                 0.28.1
huggingface_hub       1.4.1
idna                  3.11
Jinja2                3.1.6
jiter                 0.13.0
joblib                1.5.3
markdown-it-py        4.0.0
MarkupSafe            3.0.3
mdurl                 0.1.2
mpmath                1.3.0
murmurhash            1.0.15
networkx              3.6.1
numpy                 2.4.2
openai                2.20.0
packaging             26.0
pandas                3.0.0
pillow                12.1.1
pip                   26.0.1
preshed               3.0.12
pydantic              2.12.5
pydantic_core         2.41.5
Pygments              2.19.2
pytesseract           0.3.13
python-dateutil       2.9.0.post0
PyYAML                6.0.3
RapidFuzz             3.14.3
regex                 2026.1.15
requests              2.32.5
rich                  14.3.2
safetensors           0.7.0
scikit-learn          1.8.0
scipy                 1.17.0
sentence-transformers 5.2.2
setuptools            82.0.0
shellingham           1.5.4
six                   1.17.0
smart_open            7.5.0
sniffio               1.3.1
spacy                 3.8.11
spacy-legacy          3.0.12
spacy-loggers         1.0.5
srsly                 2.5.2
starlette             0.52.1
sympy                 1.14.0
thinc                 8.3.10
threadpoolctl         3.6.0
tokenizers            0.22.2
torch                 2.10.0
tqdm                  4.67.3
transformers          5.1.0
typer                 0.23.0
typer-slim            0.23.0
typing_extensions     4.15.0
typing-inspection     0.4.2
tzdata                2025.3
urllib3               2.6.3
uvicorn               0.40.0
wasabi                1.1.3
weasel                0.4.3
wheel                 0.46.3
wrapt                 2.1.1